+++ simple/library/include/visual-editor.php
@@ -2,11 +2,11 @@
 /**
  * Visual editor addons
  *
- * @package     WebMan WordPress Theme Framework
+ * @package     WebMan WordPress Theme Framework (Simple)
  * @subpackage  Visual Editor
  *
  * @since    1.0
- * @version  1.0
+ * @version  1.0.8
  *
  * Contents:
  *
@@ -24,7 +24,7 @@
 
 	// Visual Editor class
 
-		locate_template( {%= prefix_constant %}_LIBRARY_DIR . 'inc/classes/class-visual-editor.php', true );
+		locate_template( {%= prefix_constant %}_LIBRARY_DIR . 'include/classes/class-visual-editor.php', true );